requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
absl-py=1.1.0=pypi_0
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py37h540881e_2
asgiref=3.5.2=pypi_0
attrs=21.4.0=pyhd8ed1ab_0
autobahn=22.7.1=pypi_0
automat=20.2.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.11.1=pyha770c72_0
blas=1.0=mkl
bleach=5.0.0=pyhd8ed1ab_0
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.6.15=ha878542_0
cachetools=5.2.0=pypi_0
certifi=2022.6.15=py37h89c1867_0
cffi=1.15.0=py37h036bc23_0
channels=3.0.5=pypi_0
charset-normalizer=2.0.12=pypi_0
clip=1.0=pypi_0
cloudpickle=2.1.0=pypi_0
constantly=15.1.0=pypi_0
cryptography=37.0.4=pypi_0
cudatoolkit=10.2.89=h8f6ccaa_10
cycler=0.11.0=pyhd8ed1ab_0
cython=0.29.20=pypi_0
daphne=3.0.2=pypi_0
dbus=1.13.6=h5008d03_3
debugpy=1.6.0=py37hd23a5d3_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
detectron2=0.1.3+cu101=pypi_0
django=3.2.15=pypi_0
djangorestframework=3.13.1=pypi_0
einops=0.3.2=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
expat=2.4.8=h27087fc_0
ffmpeg=4.3=hf484d3e_0
filelock=3.7.1=pypi_0
flit-core=3.7.1=pyhd8ed1ab_0
fontconfig=2.14.0=h8e229c2_0
freetype=2.10.4=h0708190_1
ftfy=6.1.1=pypi_0
future=0.18.2=pypi_0
fvcore=0.1.5.post20220512=pypi_0
gettext=0.19.8.1=h73d1719_1008
glib=2.70.2=h780b84a_4
glib-tools=2.70.2=h780b84a_4
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=2.9.1=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.47.0=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=h28cd5cc_2
huggingface-hub=0.8.1=pypi_0
hyperlink=21.0.0=pypi_0
icu=58.2=hf484d3e_1000
idna=3.3=pypi_0
importlib-metadata=4.11.3=py37h89c1867_1
importlib_resources=5.7.1=pyhd8ed1ab_0
incremental=21.3.0=pypi_0
iopath=0.1.10=pypi_0
ipykernel=6.13.0=py37h25bab4e_0
ipython=7.32.0=py37h89c1867_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.7.0=pyhd8ed1ab_0
jedi=0.18.1=py37h89c1867_1
jinja2=3.1.1=pyhd8ed1ab_0
jpeg=9b=h024ee3a_2
jsonschema=4.4.0=pyhd8ed1ab_0
jupyter=1.0.0=py37h89c1867_7
jupyter_client=7.2.2=pyhd8ed1ab_1
jupyter_console=6.4.3=pyhd8ed1ab_0
jupyter_core=4.9.2=py37h89c1867_0
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_widgets=1.1.0=pyhd8ed1ab_0
kaolin=0.9.1=pypi_0
kiwisolver=1.4.2=py37h7cecad7_1
lame=3.100=h7f98852_1001
ld_impl_linux-64=2.36.1=hea4e1c9_2
libffi=3.4.2=h7f98852_5
libgcc-ng=11.2.0=h1d223b6_15
libglib=2.70.2=h174f98d_4
libiconv=1.16=h516909a_0
libnsl=2.0.0=h7f98852_0
libpng=1.6.37=h21135ba_2
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=he4da1e4_15
libtiff=4.2.0=h85742a9_0
libuuid=2.32.1=h7f98852_1000
libuv=1.43.0=h7f98852_0
libwebp-base=1.2.2=h7f98852_1
libxcb=1.13=h7f98852_1004
libxml2=2.9.9=h13577e0_2
libzlib=1.2.11=h166bdaf_1014
llvm-openmp=13.0.1=he0ac6c6_1
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.3.7=pypi_0
markupsafe=2.1.1=py37h540881e_1
matplotlib=3.4.2=py37h89c1867_0
matplotlib-base=3.4.2=py37h1058ff1_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mistune=0.8.4=py37h5e8e339_1005
mkl=2021.4.0=h8d4b97c_729
mkl-service=2.4.0=py37h402132d_0
mkl_fft=1.3.1=py37h3e078e5_1
mkl_random=1.2.2=py37h219a48f_0
mock=4.0.3=pypi_0
nbclient=0.6.0=pyhd8ed1ab_0
nbconvert=6.5.0=pyhd8ed1ab_0
nbconvert-core=6.5.0=pyhd8ed1ab_0
nbconvert-pandoc=6.5.0=pyhd8ed1ab_0
nbformat=5.3.0=pyhd8ed1ab_0
ncurses=6.3=h27087fc_1
nest-asyncio=1.5.5=pyhd8ed1ab_0
nettle=3.6=he412f7d_0
ninja=1.10.2=h4bd325d_1
notebook=6.4.11=pyha770c72_0
numpy=1.21.5=py37he7a7128_1
numpy-base=1.21.5=py37hf524024_1
oauthlib=3.2.0=pypi_0
olefile=0.46=pyh9f0ad1d_1
opencv-python=4.6.0.66=pypi_0
openh264=2.1.1=h780b84a_0
openssl=1.1.1n=h166bdaf_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.3.5=pypi_0
pandoc=2.18=ha770c72_0
pandocfilters=1.5.0=pyhd8ed1ab_0
panopticapi=0.1=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pcre=8.45=h9c3ff4c_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=9.1.0=pypi_0
pip=22.2.2=pyhd8ed1ab_0
portalocker=2.5.1=pypi_0
prometheus_client=0.14.1=pyhd8ed1ab_0
prompt-toolkit=3.0.29=pyha770c72_0
prompt_toolkit=3.0.29=hd8ed1ab_0
protobuf=3.19.4=pypi_0
psutil=5.9.0=py37h540881e_1
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.4=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pydot=1.4.2=pypi_0
pygments=2.11.2=pyhd8ed1ab_0
pymeshlab=2022.2.post2=pypi_0
pyopenssl=22.0.0=pypi_0
pyparsing=3.0.8=pyhd8ed1ab_0
pyqt=5.9.2=py37hcca6a23_4
pyrsistent=0.18.1=py37h540881e_1
python=3.7.12=hb7a2778_100_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-fastjsonschema=2.15.3=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
pytorch=1.9.0=py3.7_cuda10.2_cudnn7.6.5_0
pytorch3d=0.3.0=pypi_0
pytz=2022.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=22.3.0=py37h0c0c2a8_2
qt=5.9.7=h5867ecd_1
qtconsole=5.3.0=pyhd8ed1ab_0
qtconsole-base=5.3.0=pyhd8ed1ab_0
qtpy=2.0.1=pyhd8ed1ab_0
readline=8.1=h46c0cb4_0
regex=2022.3.15=pypi_0
requests=2.28.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.8=pypi_0
scipy=1.5.2=pypi_0
seaborn=0.11.2=pypi_0
send2trash=1.8.0=pyhd8ed1ab_0
service-identity=21.1.0=pypi_0
setuptools=62.1.0=py37h89c1867_0
sip=4.19.8=py37hf484d3e_1000
siren-pytorch=0.1.5=pypi_0
six=1.16.0=pyh6c4a22f_0
soupsieve=2.3.1=pyhd8ed1ab_0
sqlite=3.38.2=h4ff8645_0
sqlparse=0.4.2=pypi_0
tabulate=0.8.10=pypi_0
tbb=2021.5.0=h924138e_1
tensorboard=2.9.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.13.3=py37h89c1867_1
tinycss2=1.1.1=pyhd8ed1ab_0
tk=8.6.12=h27826a3_0
tokenizers=0.12.1=pypi_0
torchaudio=0.9.0=py37
torchtext=0.10.0=pypi_0
torchvision=0.10.0=py37_cu102
tornado=6.1=py37h540881e_3
tqdm=4.64.0=pypi_0
traitlets=5.1.1=pyhd8ed1ab_0
transformers=4.20.1=pypi_0
trimesh=3.9.33=pyh6c4a22f_0
twisted=22.4.0=pypi_0
txaio=22.2.1=pypi_0
typing-extensions=4.2.0=hd8ed1ab_1
typing_extensions=4.2.0=pyha770c72_1
urllib3=1.26.9=pypi_0
usd-core=21.8=pypi_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
werkzeug=2.1.2=pypi_0
wheel=0.37.1=pyhd8ed1ab_0
widgetsnbextension=3.6.0=py37h89c1867_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xz=5.2.5=h516909a_1
yacs=0.1.8=pypi_0
zeromq=4.3.4=h9c3ff4c_1
zipp=3.8.0=pyhd8ed1ab_0
zlib=1.2.11=h166bdaf_1014
zope-interface=5.4.0=pypi_0
zstd=1.4.9=ha95c52a_0
